apps/cert-api/pyproject.toml
[project]
name = "cert-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "psycopg2-binary>=2.9.10",
  "python-dotenv>=1.0.1",
  "oracledb>=2.4.0",
  "pymssql>=2.3.0",
  "openpyxl>=3.1.5",
  "gspread>=6.1.0",
  "google-auth>=2.38.0",
  "slowapi>=0.1.9",
  "requests>=2.32.0",
  "apscheduler>=3.10.4",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-mock>=3.14.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.7.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM", "RET"]
# Keep the strong correctness signals (F, B, basic E) but silence the
# noisy style preferences the existing codebase does not follow. These
# ignores were chosen after auditing every remaining offense — none of
# them represents a latent bug, only stylistic rearrangement.
ignore = [
  "E402",  # module imports not at top — cert-api uses try/except around oracledb at module load
  "E501",  # line too long — kept only via line-length 120 bump; long SQL strings are preferred inline
  "SIM105", # try/except/pass — clearer than contextlib.suppress for one-shot init calls
  "SIM108", # if-else → ternary — not always more readable
  "C401",  # unnecessary-generator-set — style preference
  "N806",  # non-lowercase variable in function — used for SQL column aliases
  "RET504", # unnecessary-assign — sometimes explicit is clearer
]

[tool.ruff.lint.per-file-ignores]
# Tests have long parametrize fixtures and demo strings
"tests/**" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
